requirements.txt
picamera
systemd
tweepy
paho-mqtt